refactor: remove unit directory
