+++ tests/SequenceValueOutOfBoundsOnUpdateTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Gurgentil\LaravelEloquentSequencer\Tests\Unit;
+namespace Gurgentil\LaravelEloquentSequencer\Tests;
 
 use Facades\Gurgentil\LaravelEloquentSequencer\Tests\Factories\Factory;
 use Gurgentil\LaravelEloquentSequencer\Exceptions\SequenceValueOutOfBoundsException;
-use Gurgentil\LaravelEloquentSequencer\Tests\TestCase;
 
 class SequenceValueOutOfBoundsOnUpdateTest extends TestCase
 {
