+++ tests/UpdateObjectWithSequenceValueLessThanCurrentSequenceValueTest.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Gurgentil\LaravelEloquentSequencer\Tests\Unit;
+namespace Gurgentil\LaravelEloquentSequencer\Tests;
 
 use Facades\Gurgentil\LaravelEloquentSequencer\Tests\Factories\Factory;
-use Gurgentil\LaravelEloquentSequencer\Tests\TestCase;
 
 class UpdateObjectWithSequenceValueLessThanCurrentSequenceValueTest extends TestCase
 {
